Rebase refactor. Let Items required changes
Rebase was originally written with the intention of being a public method.
However, as its only public use would relate to a change in lower bound it was just called internally.
The child function recursiveRebase() was also being called elsewhere in the class.
This commit sees Rebase refactored to remove the complexity introduced by premature optimisation
and changes the signature. All internal calls now go through Rebase and not recursiveRebase.

This change necessitated some changes to the Push() method and allowed for a more sophisticated logic when
adding one type of array to another.

The let Items property was also able to be made more concise with this Rebase refactor.