fix: don't iterate index - just cursor
Closes #3

This commit fixes a bug in `ParseCSV` where the index `j` was iterated
after a parsed fragement was stored, causing a delimiter to be missed
under certain conditions (e.g. when formatted as `" ,"` rather than `,`
`","` or `" , "`).

This iteration was initially added to ensure the cursor moved by two chars
so that the next string fragment started at the correct location. This
is fixed by removing the iteration of `j` previously on `BetterArray.cls`
ln 2030 and by manually moving the cursor two spaces using `cursor = j + 2`