fix: test for nullstring in UnquoteString
Closes #4

This commit adds a test trap cases where a null string is passed to
the UnquoteString function (as used by the CSV parsing methods).

Where a null/zero length string was passed to this function a runtime
error would occur as the function tries to access the first and last
characters indicies of the string.

This includes an accompanying unit test.