feat: ToExcelRange truncate where exceeds cell cnt
In cases where the number of cells or columns would exceed the available
cells in a worksheet, ToExcelRange will now truncate the array to fit
to the available space without warning.

The default behaviour or one-dimension arrays is now to print across
rows. If the Transpose param is true it will print across columns. This
is the opposite of the default behaviour in previous versions.